pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "api-debug-env"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "websockets",
    "openai",
]

[project.scripts]
server = "server.app:main"

[tool.hatch.build.targets.wheel]
packages = ["."]
